Implementado provider de conversão

+++ desafiobtg/lib/modules/convert/infra/models/convert_model.dart
@@ -10,5 +10,5 @@ class ConvertModel extends Convert{
     return '{ ${this.from}, ${this.to} }';
   }
 
-  static ConvertModel toConvert(Map<String, dynamic> map) => map == null? null: ConvertModel(from: map.values.first, to: map.values.last);
+  static ConvertModel toConvert(Map<String, dynamic> map) => map == null? null: ConvertModel(from: map.values.first.toDouble(), to: map.values.last.toDouble());
 }
